agent/pyproject.toml
[project]
name = "guia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.0",
    "fastmcp>=2.14.1",
    "markdown>=3.10",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guia"]
